[#80086] 케이웨더 shko API 데이터 삽입 완료

+++ src/main/java/com/example/realpilot/externalApiModel/kweatherAmPm7/DayOfAmPm7.java
@@ -11,10 +11,10 @@ public class DayOfAmPm7 {
     private String pmicon;
     private String amwtext;
     private String pmwtext;
-    private Integer ammintemp;
-    private Integer pmmintemp;
-    private Integer ammaxtemp;
-    private Integer pmmaxtemp;
-    private Integer amrainp;
-    private Integer pmrainp;
+    private Float ammintemp;
+    private Float pmmintemp;
+    private Float ammaxtemp;
+    private Float pmmaxtemp;
+    private Float amrainp;
+    private Float pmrainp;
 }
